:hammer: Replace Jsoup with ksoup to enable multi-platform compatibility
