+++ app/build.gradle.kts
@@ -93,7 +93,6 @@ kotlin {
             implementation(libs.jna)
             implementation(libs.jna.platform)
             implementation(libs.jnativehook)
-            implementation(libs.jsoup)
             implementation(libs.ktor.client.content.negotiation)
             implementation(libs.ktor.client.core)
             implementation(libs.ktor.client.cio)
@@ -135,6 +134,7 @@ kotlin {
             implementation(libs.kotlinx.coroutines.swing)
             implementation(libs.kotlinx.datetime)
             implementation(libs.kotlinx.serialization.json)
+            implementation(libs.ksoup)
             implementation(libs.lifecycle.common)
             implementation(libs.lifecycle.runtime)
             implementation(libs.lifecycle.runtime.compose)
